+++ tailwindcss-addons/src/utilities/index.cjs
@@ -2,7 +2,7 @@ const dir = require('./dir.cjs')
 const drag = require('./drag.cjs')
 const flip = require('./flip.cjs')
 const hideShow = require('./hideShow.cjs')
-const inputHideArrows = require('./inputHideArrows.cjs')
+const inputResets = require('./inputResets.cjs')
 const insetCenter = require('./insetCenter.cjs')
 const overflowUnset = require('./overflowUnset.cjs')
 const shortPlacements = require('./shortPlacements.cjs')
@@ -12,7 +12,7 @@ module.exports = {
 	drag,
 	flip,
 	hideShow,
-	inputHideArrows,
+	inputResets,
 	insetCenter,
 	overflowUnset,
 	shortPlacements,