+++ tailwindcss-addons/src/utilities/inputResets.cjs
@@ -3,7 +3,7 @@ const helpers = require('../lib/helpers.cjs')
 
 module.exports = plugin(({ addUtilities }) => {
 	addUtilities({
-		'.input-hide-arrows': {
+		'.number-input-reset': {
 			'&::-webkit-outer-spin-button, &::-webkit-inner-spin-button': {
 				// Using `display: none` crashes Chrome on hover
 				'-webkit-appearance': 'none',
@@ -13,6 +13,11 @@ module.exports = plugin(({ addUtilities }) => {
 			'&[type=number]': {
 				// Firefox
 				'-moz-appearance': 'textfield',
+			},
+		},
+		'.search-input-reset': {
+			'&[type=search]::-webkit-search-decoration, &[type=search]::-webkit-search-cancel-button, &[type=search]::-webkit-search-results-button, &[type=search]::-webkit-search-results-decoration': {
+				'-webkit-appearance': 'none',
 			},
 		},
 	})